Split consent from introduction.
Changed icon to be generic questionaire.
