+++ package.json
@@ -75,6 +75,7 @@
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.25.1",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "flow-bin": "^0.158.0",
     "gh-pages": "^3.2.3",
     "jest-watch-typeahead": "^0.6.5",
