+++ src/components/Consent.jsx
@@ -3,33 +3,16 @@ import { useNavigate } from "react-router-dom";
 import { DateTime } from "luxon";
 import {
   Grid,
-  TextField,
   Button,
   Box,
   Typography,
   ThemeProvider,
 } from "@material-ui/core";
-import { makeStyles } from "@material-ui/core/styles";
-import InputLabel from "@material-ui/core/InputLabel";
-import FormHelperText from "@material-ui/core/FormHelperText";
-import FormControl from "@material-ui/core/FormControl";
-import NativeSelect from "@material-ui/core/NativeSelect";
 import { useDispatch } from "react-redux";
-import * as countries from "./countries.json";
 import { dateToState } from "../features/ConversionUtil";
-import { consentShown, setDemographic } from "../features/questionSlice";
+import { consentShown, consentCompleted } from "../features/questionSlice";
 import { styles, theme } from "./ScreenHelper";
 import "../App.css";
-
-const useStyles = makeStyles((theme) => ({
-  formControl: {
-    margin: theme.spacing(1),
-    minWidth: 120,
-  },
-  selectEmpty: {
-    marginTop: theme.spacing(2),
-  },
-}));
 
 export function Consent() {
   const dispatch = useDispatch();
@@ -38,115 +21,41 @@ export function Consent() {
     dispatch(consentShown(dateToState(DateTime.utc())));
   }, []);
 
-  const classes = useStyles();
-  const [disableSubmit, setDisableSubmit] = React.useState(true);
-  const [disableSelfDescribe, setDisableSelfDescribe] = React.useState(true);
-  const [country, setCountry] = React.useState("");
-  const [visFamiliarity, setVisFamiliarity] = React.useState("");
-  const [age, setAge] = React.useState("");
-  const [gender, setGender] = React.useState("");
-  const [selfDescribeGender, setSelfDescribeGender] = React.useState("");
-  const [profession, setProfession] = React.useState("");
-
   const navigate = useNavigate();
-
-  const checkEnableSubmit = () => {
-    if (
-      country &&
-      country.length > 1 &&
-      visFamiliarity &&
-      visFamiliarity.length > 0 &&
-      age &&
-      age.length > 1 &&
-      gender &&
-      gender.length > 0 &&
-      (gender !== "self-describe" ||
-        (selfDescribeGender && selfDescribeGender.length > 0)) &&
-      profession &&
-      profession.length > 0
-    ) {
-      setDisableSubmit(false);
-    } else {
-      setDisableSubmit(true);
-    }
-  };
-
-  const handleFieldChange = (event, setter) => {
-    // reject negative ages.
-    // if (event.target.value <= 0) {
-    //   event.target.value = "";
-    // } else {
-    setter(event.target.value);
-    checkEnableSubmit();
-    if (event.target.value === "self-describe") {
-      setDisableSelfDescribe(false);
-    } else if (
-      ["female", "male", "transgender", "non-binary", "intersex"].includes(
-        event.target.value
-      )
-    ) {
-      setDisableSelfDescribe(true);
-    }
-    //    }
-  };
-
-  const vizFamiliarityLevel = [
-    {
-      id: 1,
-      name: " 1 (I don't use charts and data visualizations)",
-      value: 1,
-    },
-    { id: 2, name: " 2", value: 2 },
-    { id: 3, name: " 3", value: 3 },
-    { id: 4, name: " 4", value: 4 },
-    { id: 5, name: " 5", value: 5 },
-    { id: 6, name: " 6", value: 6 },
-    {
-      id: 7,
-      name: " 7 (I use charts and data visualizations everyday)",
-      value: 7,
-    },
-  ];
 
   const ConsentTextEn = () => {
     return (
       <React.Fragment>
-        <p>
-          <b>The goal of this research is </b> to understand how choices between
-          receiving two amounts of money, one sooner and the other later, are
-          affected by the way they are presented.
-        </p>
-        <p>
-          <b>Procedure: </b>You will be presented with a series of worded
-          questions, or barcharts or calendar visualizations of smaller earlier
-          and larger later amounts of money.
-          <b>You will choose</b> either the earlier or later amount. You may
-          also be asked to interact with the visualization <b>to adjust</b> the
-          barchart height of the earlier or later amount by dragging the mouse
-          vertically.{" "}
-        </p>
+        <Typography paragraph>
+          <b>The goal of this research is </b> to study how people make choices
+          about receiving money.
+        </Typography>
+        <Typography paragraph>
+          <b>Procedure: </b>You will be presented with a series of choices about
+          receiving money at different points in time.
+          <b>You will choose</b> either the earlier or later amount.{" "}
+        </Typography>
         {/* For all parts, do NOT provide numerical answers, always think of words and expressions instead!  */}
-        <p>
-          <b>What you will get: </b> You will learn about how visualization of
-          money amounts effects your decision to choose earlier or later
-          amounts.
-        </p>
-        <p>
+        <Typography paragraph>
+          <b>What you will get: </b> You will learn more about the goal of this
+          reasearch and how people make decisions about money.
+        </Typography>
+        <Typography paragraph>
           <b>Duration: </b> The experiment will take you about 10 minutes.
-        </p>
-        <p>
+        </Typography>
+        <Typography paragraph>
           <b>Privacy: </b> Your participation is anonymous and your responses
           cannot be used to identify you.
-        </p>
-        <p>
+        </Typography>
+        <Typography paragraph>
           <b>Record keeping: </b>Records of your participation will be held
           confidential so far as permitted by law. However, the study
           investigators and, under certain circumstances, the Worcester
           Polytechnic Institute Institutional Review Board (WPI IRB) will be
           able to inspect and have access to this data. Any publication or
           presentation of the data will not identify you.
-        </p>
-        <p>
+        </Typography>
+        <Typography paragraph>
           <b>Contact info: </b>
           {[
             { name: "Prof. Daniel Reichman", email: "dreichman@wpi.edi" },
@@ -166,8 +75,8 @@ export function Consent() {
               </span>
             );
           })}
-        </p>
-        <p>
+        </Typography>
+        <Typography paragraph>
           <b>Your participation in this research is voluntary. </b> Your refusal
           to participate will not result in any penalty to you or any loss of
           benefits to which you may otherwise be entitled. You may decide to
@@ -175,14 +84,18 @@ export function Consent() {
           of other benefits. The project investigators retain the right to
           cancel or postpone the experimental procedures at any time they see
           fit.
-        </p>
-        <p>
+        </Typography>
+        <Typography paragraph>
           For more information about your rights as a research participant, or
-          any concerns related to this study, please contact the IRB Chair
-          (Professor Kent Rissmiller, Tel. 508-831-5019, Email: kjr@wpi.edu) and
-          the Human Protection Administrator (Gabriel Johnson, Tel.
-          508-831-4989, Email: gjohnson@wpi.edu).
-        </p>
+          any concerns related to this study, please contact the:<br></br>
+          IRB Chair Professor Kent Rissmiller<br></br>
+          Tel. 508-831-5019<br></br>
+          Email: kjr@wpi.edu<br></br>
+          and the<br></br>
+          Human Protection Administrator Gabriel Johnson<br></br>
+          Tel. 508-831-4989<br></br>
+          Email: gjohnson@wpi.edu
+        </Typography>
       </React.Fragment>
     );
   };
@@ -193,14 +106,14 @@ export function Consent() {
         <Box height="25%" alignItems="center">
           <img
             style={{ maxHeight: "240px" }}
-            src="bar-chart.png"
-            alt="$100 bills on calendar"
+            src="generic-questionaire-icon.svg"
+            alt="Questionaire image."
           ></img>
         </Box>{" "}
         <Grid container style={styles.root} justifyContent="center">
           <Grid item xs={12}>
             <Typography variant="h5">
-              <b>Money Earlier or Later?</b>
+              <b>Choices About Money?</b>
               <br />
             </Typography>
             <hr
@@ -212,9 +125,9 @@ export function Consent() {
             />
             <Typography>
               We often have to make choices about receiveing rewards at
-              different point in time.
+              different points in time.
             </Typography>
-            <Typography>
+            <Typography paragraph>
               <br />
               <i>
                 {" "}
@@ -237,132 +150,9 @@ export function Consent() {
               <ConsentTextEn />
             </div>
 
-            <Typography>
-              By clicking &ldquo;Next&ldquo;, you agree to participate. Before
-              we begin, please enter your:
+            <Typography paragraph>
+              By clicking &ldquo;Next&ldquo;, you agree to participate.
             </Typography>
-          </Grid>
-          <Grid item xs={12}>
-            <FormControl
-              className={classes.formControl}
-              required
-              style={{ maxWidth: 230, marginRight: 20 }}
-            >
-              <InputLabel htmlFor="country-select-helper">
-                Country of residence
-              </InputLabel>
-              <NativeSelect
-                value={country}
-                onChange={(event) => {
-                  handleFieldChange(event, setCountry);
-                }}
-                inputProps={{
-                  name: "country-of-origin",
-                  id: "country-select-helper",
-                }}
-              >
-                <option> </option>
-                {countries.default.map((option) => (
-                  <option
-                    key={option.alpha3}
-                    id={option.alpha3}
-                    value={option.alpha3}
-                  >
-                    {option.name}
-                  </option>
-                ))}
-              </NativeSelect>
-              <FormHelperText>The country you are living in now</FormHelperText>
-            </FormControl>
-            <FormControl
-              className={classes.formControl}
-              required
-              style={{ maxWidth: 230, marginRight: 20 }}
-            >
-              <NativeSelect
-                value={visFamiliarity}
-                onChange={(event) => {
-                  handleFieldChange(event, setVisFamiliarity);
-                }}
-                name="familiarity-with-viz"
-                className={classes.selectEmpty}
-                inputProps={{ "aria-label": "Datavis experience" }}
-              >
-                <option> </option>
-                {vizFamiliarityLevel.map((option) => (
-                  <option key={option.id} value={option.id}>
-                    {option["name"]}
-                  </option>
-                ))}
-              </NativeSelect>
-              <FormHelperText>
-                Your experience with data visualizations and charts
-              </FormHelperText>
-            </FormControl>
-          </Grid>
-          <Grid item xs={12} style={{ margin: 0 }}>
-            <TextField
-              required
-              className={classes.formControl}
-              label="Age"
-              type="number"
-              id="Age"
-              onChange={(event) => {
-                handleFieldChange(event, setAge);
-              }}
-            />
-            <label style={{ marginRight: 20 }}> </label>
-            <FormControl
-              className={classes.formControl}
-              required
-              style={{ maxWidth: 230, marginRight: 20 }}
-            >
-              <InputLabel htmlFor="gender-select-helper">Gender</InputLabel>
-              <NativeSelect
-                value={gender}
-                onChange={(event) => {
-                  handleFieldChange(event, setGender);
-                }}
-                inputProps={{
-                  name: "gender",
-                  id: "gender-select-helper",
-                }}
-              >
-                <option> </option>
-                {[
-                  { value: "female", text: "Female" },
-                  { value: "male", text: "Male" },
-                  { value: "transgender", text: "Transgender" },
-                  { value: "non-binary", text: "Non-binary" },
-                  { value: "intersex", text: "Intersex" },
-                  { value: "self-describe", text: "Prefer to self-describe" },
-                ].map(({ value, text }) => (
-                  <option key={value} id={value} value={value}>
-                    {text}
-                  </option>
-                ))}
-              </NativeSelect>
-            </FormControl>
-            <TextField
-              required
-              className={classes.formControl}
-              label="Self Describe Gender"
-              id="Self-Describe-Gender"
-              onChange={(event) => {
-                handleFieldChange(event, setSelfDescribeGender);
-              }}
-              disabled={disableSelfDescribe}
-            />
-            <label style={{ marginLeft: 25 }}> </label>
-            <TextField
-              required
-              className={classes.formControl}
-              label="Current Profession"
-              id="Current-Profession"
-              onChange={(event) => {
-                handleFieldChange(event, setProfession);
-              }}
-            />
           </Grid>
           <Grid item xs={12} style={{ margin: 0 }}>
             <Button
@@ -372,18 +162,9 @@ export function Consent() {
               disableFocusRipple
               style={styles.button}
               onClick={() => {
-                dispatch(
-                  setDemographic({
-                    country: country,
-                    visFamiliarity: visFamiliarity,
-                    age: age,
-                    gender: gender,
-                    profession: profession,
-                  })
-                );
-                navigate("/introduction");
+                dispatch(consentCompleted(dateToState(DateTime.utc())));
+                navigate("/demographic");
               }}
-              disabled={disableSubmit}
             >
               {" "}
               Next{" "}
